# Cache reflection analysis in RowToStructBy...