@jackc 2024-04-16
Definitely a big win. Thanks!

Before:

```
jack@glados ~/dev/pgx ±8db0f280 » got -run=^$ -bench='SelectRowsSimpleAppend' -benchmem
goos: darwin
goarch: arm64
pkg: github.com/jackc/pgx/v5
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/1_rows-16         	   30272	     37969 ns/op	    1017 B/op	      28 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/10_rows-16        	   25342	     47388 ns/op	    4267 B/op	     163 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/100_rows-16       	    8980	    131713 ns/op	   36696 B/op	    1513 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/1000_rows-16      	    1963	    600679 ns/op	  361149 B/op	   15020 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/1_rows-16        	   31489	     37810 ns/op	    1032 B/op	      29 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/10_rows-16       	   25466	     46919 ns/op	    4267 B/op	     164 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/100_rows-16      	    9043	    132128 ns/op	   36697 B/op	    1514 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/1000_rows-16     	    1965	    604181 ns/op	  361151 B/op	   15021 allocs/op
PASS
ok  	github.com/jackc/pgx/v5	11.535s
```

After:

```
jack@glados ~/dev/pgx ±8f69e45a » got -run=^$ -bench='SelectRowsSimpleAppend' -benchmem
goos: darwin
goarch: arm64
pkg: github.com/jackc/pgx/v5
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/1_rows-16         	   30673	     37387 ns/op	     920 B/op	      18 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/10_rows-16        	   27057	     44155 ns/op	    3274 B/op	      63 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/100_rows-16       	   10000	    108665 ns/op	   26817 B/op	     513 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByPos/1000_rows-16      	    2806	    420271 ns/op	  262453 B/op	    5023 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/1_rows-16        	   31635	     38588 ns/op	     936 B/op	      19 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/10_rows-16       	   26431	     45148 ns/op	    3290 B/op	      64 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/100_rows-16      	   10000	    110390 ns/op	   26833 B/op	     514 allocs/op
BenchmarkSelectRowsSimpleAppendRowsRowToStructByName/1000_rows-16     	    2716	    428189 ns/op	  262469 B/op	    5024 allocs/op
PASS
ok  	github.com/jackc/pgx/v5	11.254s
```